+++ packages/orq-rc/src/lib/config.ts
@@ -64,7 +64,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.6.0-rc.18",
+  sdkVersion: "3.6.0-rc.19",
   genVersion: "2.605.0",
-  userAgent: "speakeasy-sdk/typescript 3.6.0-rc.18 2.605.0 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.6.0-rc.19 2.605.0 2.0 @orq-ai/node",
 } as const;