+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -177,7 +177,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-05-20T13:20:39.505Z",
+    "2025-05-20T13:37:33.567Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -214,7 +214,7 @@ export const CreateContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-05-20T13:20:39.505Z"))
+  updated: z.date().default(() => new Date("2025-05-20T13:37:33.567Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {