+++ packages/orq-rc/RELEASES.md
@@ -1619,3 +1619,13 @@ Based on:
 - [typescript v3.6.0-rc.18] packages/orq-rc
 ### Releases
 - [NPM v3.6.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.18 - packages/orq-rc
+
+## 2025-05-20 13:38:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.548.2 (2.605.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.6.0-rc.19] packages/orq-rc
+### Releases
+- [NPM v3.6.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/3.6.0-rc.19 - packages/orq-rc